# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PubKeyMail API
  version: 1.0.0
  description: >-
    Blockchain-based email service with wallet authentication.


    ## Authentication

    PubKeyMail uses wallet-based authentication. Users sign a challenge message 

    with their blockchain wallet and receive a JWT token for subsequent
    requests.


    ## Case Sensitivity

    **CRITICAL**: All blockchain addresses are case-sensitive. Different casing 

    represents different addresses.
  contact:
    name: PubKeyMail Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pubkeymail.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/verify endpoint
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: array
              items: {}
          required:
            - code
            - message
      required:
        - error
    ChallengeRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
          example: GNa9E2dWP8e2fg23pRJoiCvJm5cbzomm94YzotpcRh7A
          description: Blockchain wallet address
        blockchain:
          type: string
          enum:
            - solana
            - ethereum
            - polygon
          default: solana
      required:
        - address
    ChallengeResponse:
      type: object
      properties:
        challenge:
          type: string
          example: "Sign this message to authenticate with PubKeyMail: abc123..."
          description: Message to sign with wallet
        nonce:
          type: string
          example: abc123def456
          description: Unique nonce for this challenge
        expiresAt:
          type: string
          format: date-time
          example: 2024-01-01T12:05:00Z
      required:
        - challenge
        - nonce
        - expiresAt
    VerifyRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
          example: GNa9E2dWP8e2fg23pRJoiCvJm5cbzomm94YzotpcRh7A
        blockchain:
          type: string
          enum:
            - solana
            - ethereum
            - polygon
          default: solana
        signature:
          type: string
          minLength: 1
          example: 3xJ9fK...
          description: Base58 or Base64 encoded signature
        nonce:
          type: string
          minLength: 1
          example: abc123def456
      required:
        - address
        - signature
        - nonce
    VerifyResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token for authenticated requests
        expiresIn:
          type: number
          example: 86400
          description: Token validity in seconds
        user:
          type: object
          properties:
            address:
              type: string
            blockchain:
              type: string
              enum:
                - solana
                - ethereum
                - polygon
          required:
            - address
            - blockchain
      required:
        - token
        - expiresIn
        - user
...
    SendEmailRequest:
      type: object
      properties:
        userId:
          type: number
          exclusiveMinimum: 0
        fromAddressId:
          type: number
          exclusiveMinimum: 0
        toAddress:
          type: string
          format: email
          example: recipient@example.com
        subject:
          type: string
          minLength: 1
          maxLength: 998
        bodyText:
          type: string
        bodyHtml:
          type: string
        replyTo:
          type: string
          format: email
      required:
        - userId
        - fromAddressId
        - toAddress
        - subject
    SendEmailResponse:
      type: object
      properties:
        success:
          type: boolean
        messageId:
          type: string
        emailId:
          type: string
          format: uuid
        details:
          type: object
          properties:
            provider:
              type: string
            deliveryStatus:
              type: string
              enum:
                - pending
                - sent
                - delivered
                - failed
                - bounced
            sentAt:
              type: string
              format: date-time
          required:
            - provider
            - deliveryStatus
            - sentAt
      required:
        - success
        - emailId
    SentEmail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from:
          type: string
        to:
          type: string
        subject:
          type: string
        sentAt:
          type: string
          format: date-time
        deliveryStatus:
          type: string
          enum:
            - pending
            - sent
            - delivered
            - failed
            - bounced
      required:
        - id
        - from
        - to
        - subject
        - sentAt
        - deliveryStatus
    SentEmailDetail:
      allOf:
        - $ref: "#/components/schemas/SentEmail"
        - type: object
          properties:
            bodyText:
              type: string
            bodyHtml:
              type: string
            messageId:
              type: string
    FromAddress:
      type: object
      properties:
        id:
          type: number
        address:
          type: string
        email:
          type: string
          format: email
        blockchain:
          type: string
          enum:
            - solana
            - ethereum
            - polygon
      required:
        - id
        - address
        - email
        - blockchain
...
paths:
  /api/v1/auth/challenge:
    post:
      tags:
        - Auth
      summary: Request authentication challenge
      description: Request a challenge message to sign with your wallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeRequest"
      responses:
        "200":
          description: Challenge generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          description: Validation error or unsupported blockchain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/verify:
    post:
      tags:
        - Auth
      summary: Verify wallet signature
      description: Verify a signed challenge and receive a JWT token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/emails/send:
    post:
      tags:
        - Emails
      summary: Send an email
      description: Send an email (requires paid subscription)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailRequest"
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendEmailResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Subscription required or no permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User or address not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/emails/sent:
    get:
      tags:
        - Emails
      summary: Get sent emails
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: number
            exclusiveMinimum: 0
          required: true
          name: userId
          in: query
        - schema:
            type: number
            minimum: 1
            maximum: 100
            default: 50
          required: false
          name: limit
          in: query
        - schema:
            type:
              - number
              - "null"
            minimum: 0
            default: 0
          required: false
          name: offset
          in: query
      responses:
        "200":
          description: List of sent emails
          content:
            application/json:
              schema:
                type: object
                properties:
                  emails:
                    type: array
                    items:
                      $ref: "#/components/schemas/SentEmail"
                  total:
                    type: number
                  limit:
                    type: number
                  offset:
                    type: number
                required:
                  - emails
                  - total
                  - limit
                  - offset
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/emails/from-addresses:
    get:
      tags:
        - Emails
      summary: Get available from addresses
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: number
            exclusiveMinimum: 0
          required: true
          name: userId
          in: query
      responses:
        "200":
          description: List of from addresses
          content:
            application/json:
              schema:
                type: object
                properties:
                  addresses:
                    type: array
                    items:
                      $ref: "#/components/schemas/FromAddress"
                required:
                  - addresses
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/emails/{emailId}:
    get:
      tags:
        - Emails
      summary: Get sent email by ID
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: emailId
          in: path
        - schema:
            type: number
            exclusiveMinimum: 0
          required: true
          name: userId
          in: query
      responses:
        "200":
          description: Email details
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    $ref: "#/components/schemas/SentEmailDetail"
                required:
                  - email
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Email not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...